Modified ui. Added functionalities
Now in the ui is possible do select the user or the administrator. For
the user is sufficient to put the nickname, to access as administrator
it is necessary to put nickname="admin" and password="admin".
It's now possible to create new playlist in the Playlists page of the
user.